requirements.txt
fastapi
uvicorn
beanie
motor
pydantic
python-dotenv
pycryptodome
python-jose
